+++ lib/src/state_machine/mp_th2_file_edit_state_machine/mp_th2_file_edit_state.dart
@@ -148,11 +148,15 @@ abstract class MPTH2FileEditState {
 
   void onTertiaryButtonScroll(PointerScrollEvent event) {}
 
+  void onDeselectAll() {}
+
   void onKeyDownEvent(KeyDownEvent event) {}
 
   void onKeyRepeatEvent(KeyRepeatEvent event) {}
 
   void onKeyUpEvent(KeyUpEvent event) {}
+
+  void onSelectAll() {}
 
   void onStateEnter(MPTH2FileEditState previousState) {}